chore: split database traits
