+++ crate/server/src/core/implementation.rs
@@ -4,7 +4,7 @@ use tracing::{debug, instrument, trace};
 
 use crate::{
     config::{DbParams, ServerParams},
-    database::{Database, Redis},
+    database::{DatabaseTraits, Redis},
     error::result::FResult,
     findex_server_bail,
     middlewares::{JwtAuthClaim, PeerCommonName},
@@ -13,12 +13,12 @@ use crate::{
 
 pub(crate) struct FindexServer {
     pub(crate) params: ServerParams,
-    pub(crate) db: Box<dyn Database + Sync + Send>,
+    pub(crate) db: Box<dyn DatabaseTraits + Sync + Send>,
 }
 
 impl FindexServer {
     pub(crate) async fn instantiate(mut shared_config: ServerParams) -> FResult<Self> {
-        let db: Box<dyn Database + Sync + Send> =
+        let db: Box<dyn DatabaseTraits + Sync + Send> =
             if let Some(mut db_params) = shared_config.db_params.as_mut() {
                 match &mut db_params {
                     DbParams::Redis(url) => Box::new(
