+++ crate/server/src/database/mod.rs
@@ -1,7 +1,7 @@
-mod database_trait;
+mod database_traits;
 mod redis;
 
-pub(crate) use database_trait::Database;
+pub(crate) use database_traits::DatabaseTraits;
 pub(crate) use redis::Redis;
 
 #[cfg(test)]
